código pronto

+++ index.js
@@ -5,7 +5,7 @@ let posicao = [];// variável para idades declarada.
 for (i = 0; i < 4; i++) {
     let idade;
     do {
-        idade = parseFloat(prompt(`Digite a idade ${i+1}`))
+        idade = parseInt(prompt(`Digite a idade ${i+1}`))
         if (isNaN(idade) || idade < 0 ) {
             alert("Idade inválida, digite novamente");
         }
